Refactor worker code, introduce RIO module
Split fastly logging, Redirection.io configuration and application to better
support other cases and later on simplify custom-caching rules applied to request
